node/demo/requirements.txt
pinata
python-dotenv
web3==6.8.0
base58
eth-event
docker>=2
structlog>=15.3
python-dateutil>=2.4